Implementations in simplemovement.c and more API
Added five methods to simplemovement, and updated the header to reflect
that some of the functions have been implemented. The header has also
been completely restructred, and should be much prettier now. Some stuff
has also been tweaked to allow for better API use; see the update on
"Improved functionality as an API" for more info on what exactly the API
changes entailed.