Updating types- Part 1
Updating some data types of variables to better comply with RobotC's
compiler. This may be in more than one installment; I can't remember if
I finished this. Exact changes detailed:
--Removed all ubytes: RobotC does not support unsigned variables, so
these are as good as normal bytes
--Removed all shorts: RobotC compiles all 'int' variables as shorts
instead, so they have been changed to int.
--Changed some bytes to int- ubytes are not supported, so the max value
of a byte it 128. For variables that this is not enough for, the type
has been changed to int/short.