Created timer.c and header
Should be used for our own timer structure. This will improve the native
RobotC timer system, because not only are we no longer limited to four
timers, but we can also change them any way we want.

Daniel, please verify that my use of pointers is correct, and add a
declaration for the Timer_t structure in the header.