Debugging Part 2- All changes listed
--Renamed the file "default config" to "default_config" as it is now
referenced in the program.
--Reverted Daniel's change to pointer arrays because it doesn't compile
and I have no idea what he was trying to do anyway
--Github thinks I moved a file from /api/ to /API/ so I guess there's
that. I can't explain it but it doesn't look to be causing any problems.
--Added some include statements because RobotC was complaining about
things being undefined.

The code currently compiles fine, so we're still good to test things. We
should start writing test programs to get things moving on Friday.