Changed compass sensor structure
Created a method for compass sensor value in sensors. This way, we can
make sure the sensor is defined before trying to read it. Also changed
some stuff around in simplemovement, including but not limited to
updating the header, deprecating functions, re-organizing functions, and
adding functionality. Also made changes to the API tutorial.