Attempt at fixing timer.c timer array
This commit is for debugging purposes only
Also renamed clearTimer to emptyTimer so that RobotC
does not complain. All edits in this commit were made
and commited from the linux command line.

An array in C is really just a pointer.

int array[10];
array[i] = 7; //equivalent of &array + i*sizeof(int) = 7;

The previous code is what the compiler will correct all uses of arrays to.
As RobotC can not have arrays as parameters to functions, changing the code
manually might allow arrays of pointers.

TODO: look up equivalent of sizeof in RobotC (if it exists)