Removed fields from the timer structure
The only field in Timer_t is now milliseconds, and any other measurement
of time can be accessed by using a function whenever it is needed. This
way will save lots of unnecessary memory from the fields, and will make
the program faster by not calculating things then not necessary.