Debugging/ reverting
Daniel, a couple lines you changed have been reverted, but I couldn't
revert the whole commit because there was other stuff that needed to
stay.
--Changed timerSet[] back to being an array of pointers. It has to be
like this because RobotC sucks and we can't actually do it any other
way. No dynamic allocation and such.
--Changed the if/else structure so that timerSet[] will actually be
defined.
--Reverted the updateTimer() argument problem. Logic tells us it should
have been done Daniel's way, but passing the address is the only thing
that makes it compile. We may fix this by having updateTimer() accept a
pointer to a pointer to a timer, instead of a pointer to a timer.