Brought NRMT up to date
If these changes break the program, use the old version that I've
included in teleop/old. It's pretty much exactly the same code, except
that it's been made more modular and it uses constants in settings.c.

If these changes work as expected, main_teleop will be deprecated, and
this file will be renamed to main_teleop.