Proposed change for exit button problem
This code will exit the program if the exit button is pressed three
times consecutively. If a mode is accidentally passed by while pressing
the orange button, you can still use the grey button to back up one or
two modes at a time. If a problem arises and the program must stop,
press the exit button three times (at any speed) to stop the program.
PROMPTEXIT is still in the enumeration and the string array, as well as
the main function loop, but has been removed from nextMode and lastMode
so it will be skipped over and never actually run.