Implemented new lift motor & UPDATES
To  @danielgrimshaw and @pranavmathur : It is very important for both of
you to read the pre-scrim text. It will serve as a planning document
because we will not see each other until after the competition.

New stuff:
--Changed 'lift' motor to 'grab' to keep the old implementations
--We will need to set all the time/speed values for the new motors
--I left all the 'lift' stuff commented, because the lift is really
functioning as a goal 'grab' and it acts as such in the code.
--Brought all current configs up to date (excluding autonomous programs)