Rewrote api/sensors.c
Now it only contains wrapper functions for Xander's driver suite. There
are also definition checks to make sure a function doesn't call a sensor
that doesn't exist. A consequence of this is that sensors will need to
be defined in the config. I will update the autopush to copy select
definitions as well.