Temporarily disabled coord system
Only a temporary change. I want to change the coordinate system to
listen to the motors, instead of having to be manually updated. I also
changed some functions to clarify their specificity to the coordinate
system (setDefaults and readValues are too vague names; they could apply
to anything). In names that I changed, the original name has been
temporarily kept as a static function.