CODE COMPILES PERFECTLY!!
Please stop blaming all the 'code being broken' on me. It was my fault
one time, and even that was minor. The rest of it was only a byproduct
of trying to dig us out of the pit that the branch fiasco put us into.

Most of the edits here are include statements. You guys could've figured
this one out just fine.
--Removed the debug include from background
--Added a proper include guard to api/coord.c
--Removed the include to coord.c from sensors.h
--Re-implemented the bool usePID to simplemovement after it was lost in
the merge clusterfuck. (This is probably the only real change in the
entire commit, and it's literally 6 lines; 9 if you count the header)
--Include api/coord.c from autoV1.c