autofixWhitespace.py now works with all comments
@ZachOhara @zachw000 @pranavmathur :

I have not yet even looked at the other problems.
The following example is showing one of the few known problems.
EX:
```
/*
multiline comment one
this code here is kept in the comment*/pretend this is valid code here /* comment 2*/ /*for
multiline comment two
*/
```

NOTES:
	* The multiline comment two will notbe indented correctly. The for at the beginning of the comment will be fed through the whitespace finding algorithm, making the tabbing thereafter be the same as though it was normal code. I AM NOT GOING TO FIX THIS, IT IS FAR TOO MUCH EFFORT FOR THE REWARD.
	* If you can see random backticks (`) and asterisks (*) that means that Github does not support markdown in its commit logs. I was hoping it would support it, but I don't think that it does.

If you, for some unknown reason, try to use this code in its current state and find a bug please add it to the currently open issue about this
script.