Please read:
@danielgrimshaw @zachw000 @ZachOhara

We should instate a new rule: NO MERGES should be made OR BRANCHES
deleted without first consulting all 4 programmers; if one of us is
working on a branch, but the code is not finished, and someone else
merges the branch without looking at the code inside, stuff gets messed
up. I'm not saying it's happened before, I'm just saying that we should
instate this rule. Please vote. A 3/4 majority is required to pass.